grr
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@54226 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostDrivers/Support/SUPLibAll.cpp
@@ -100,7 +100,7 @@ SUPDECL(uint64_t) SUPReadTscWithDelta(void)
     uint16_t    iGipCpu   = (unsigned)iCpuSet < RT_ELEMENTS(pGip->aiCpuFromCpuSetIdx)
                           ? pGip->aiCpuFromCpuSetIdx[iCpuSet] : UINT16_MAX;
 #  else
-    uint8_t     idApic    = ASMGetApicId();  /** @todo this could probably be eliminated in RC if we really wanted to... */
+    uint16_t    idApic    = ASMGetApicId();  /** @todo this could probably be eliminated in RC if we really wanted to... */
     uint16_t    iGipCpu   = (unsigned)idApic < RT_ELEMENTS(pGip->aiCpuFromApicId) /* for the future */
                           ? pGip->aiCpuFromApicId[idApic] : UINT16_MAX;
 #  endif
